+++ src/main/java/com/doctor/vampiricageing/capabilities/VampiricAgeingCapabilityManager.java
@@ -466,7 +466,7 @@ public class VampiricAgeingCapabilityManager {
             }
             int age = getAge(event.player).map(ageCap -> ageCap.getAge()).orElse(0);
             VampirePlayer.getOpt(event.player).ifPresent(vamp -> {
-                vamp.addExhaustion(CommonConfig.amountExhaustionDrainFromSources.get().get(age));
+                vamp.addExhaustion(CommonConfig.amountExhaustionDrainFromSources.get().get(age).floatValue());
             });
         }
     }
@@ -543,14 +543,14 @@ public class VampiricAgeingCapabilityManager {
 
                 if(event.getSource() != VReference.HOLY_WATER && CommonConfig.deadlySourcesFastDrainExhaustion.get() && event.getEntity() instanceof PlayerEntity) {
                     VampirePlayer.getOpt((PlayerEntity) event.getEntity()).ifPresent(vamp -> {
-                        vamp.addExhaustion(CommonConfig.amountExhaustionDrainFromSources.get().get(age));
+                        vamp.addExhaustion(CommonConfig.amountExhaustionDrainFromSources.get().get(age).floatValue());
                     });
                 }
                 event.setAmount(event.getAmount() / CommonConfig.genericVampireWeaknessReduction.get().get(age).floatValue());
             } else if(event.getSource() == DamageSource.STARVE && CommonConfig.harsherOutOfBlood.get() && age > 0) {
                 event.setAmount(event.getAmount() * age);
             } else if(event.getSource().getEntity() != null && event.getSource().getEntity().getType().is(ModTags.Entities.HUNTER) && CommonConfig.shouldAgeIncreaseHunterMobDamage.get()) {
-                event.setAmount(event.getAmount() * CommonConfig.damageMultiplierFromHunters.get().get(age));
+                event.setAmount(event.getAmount() * CommonConfig.damageMultiplierFromHunters.get().get(age).floatValue());
             } else if(!Helper.canKillVampires(event.getSource()) && event.getAmount() >= event.getEntityLiving().getHealth()  && CommonConfig.shouldOnlyDieFromKillingSources.get() && age >= CommonConfig.shouldOnlyDieFromKillingSourcesAgeRank.get()) {
                 //anyone remember witchery
                 event.getEntityLiving().setHealth(1);
@@ -563,7 +563,7 @@ public class VampiricAgeingCapabilityManager {
     public static void onHeal(LivingHealEvent event) {
         if(CommonConfig.shouldAgeAffectHealing.get() && Helper.isVampire(event.getEntity())) {
             int age = getAge(event.getEntityLiving()).map(ageCap -> ageCap.getAge()).orElse(0);
-            event.setAmount(event.getAmount() * CommonConfig.ageHealingMultiplier.get().get(age));
+            event.setAmount(event.getAmount() * CommonConfig.ageHealingMultiplier.get().get(age).floatValue());
         }
 
     }
