+++ src/main/java/com/doctor/vampiricageing/config/WerewolvesAgeingConfig.java
@@ -22,9 +22,9 @@ public class WerewolvesAgeingConfig {
     public static final ForgeConfigSpec.ConfigValue<List<? extends Integer>> devouredForNextAge;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Integer>> nourishmentMultipliers;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> biteDamageMultiplier;
-    public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> silverOilDamageMultiplier;
-    public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> formTimeMultiplier;
-    public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> healonBiteAmount;
+    public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> silverOilDamageMultiplier;
+    public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> formTimeMultiplier;
+    public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> healonBiteAmount;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> maxHealthIncrease;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> ageDamageIncrease;
 
@@ -42,9 +42,9 @@ public class WerewolvesAgeingConfig {
         exquisiteDevourWorth = COMMON_BUILDER.comment("How much a exquisite devour is worth. These are things like vampire barons").defineInRange("exquisiteDevourWorth", 8, 0, 99);
         maxHealthIncrease = COMMON_BUILDER.comment("Max Health Increase for each rank. This is addition, not multiplier based").defineList("maxHealthIncrease", Arrays.asList(0D, 2D, 2D, 4D, 4D, 6D), t -> true);
         biteDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies bite damage. Bite damage multiplier is strange and values of 0 should be treated as 1, values like 0.2 should be treated as a 1.2 multiplier").defineList("biteDamageMultiplier", Arrays.asList(0d, 0d, 0d, 0.125d, 0.25d, 0.5d), it -> true);
-        silverOilDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies damages from Silver Oil. Values are set to 1 by default as to disable it by default.").defineList("silverOilMultiplier", Arrays.asList(1f, 1f, 1f, 1f, 1f, 1f), it -> true);
-        healonBiteAmount = COMMON_BUILDER.comment("How much the player is healed for biting an entity based on rank").defineList("healonBiteAmount", Arrays.asList(0f, 0f, 0f, 1f, 2f, 2f), it -> true);
-        formTimeMultiplier = COMMON_BUILDER.comment("How much the duration of time a player can stay in werewolf form is multiplied by").defineList("formTimeMultiplier", Arrays.asList(1f, 1.5f, 2f, 3f, 4f, 5f), it -> true);
+        silverOilDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies damages from Silver Oil. Values are set to 1 by default as to disable it by default.").defineList("silverOilMultiplier", Arrays.asList(1d, 1d, 1d, 1d, 1d, 1d), it -> true);
+        healonBiteAmount = COMMON_BUILDER.comment("How much the player is healed for biting an entity based on rank").defineList("healonBiteAmount", Arrays.asList(0d, 0d, 0d, 1d, 2d, 2d), it -> true);
+        formTimeMultiplier = COMMON_BUILDER.comment("How much the duration of time a player can stay in werewolf form is multiplied by").defineList("formTimeMultiplier", Arrays.asList(1d, 1.5d, 2d, 3d, 4d, 5d), it -> true);
         ageBuffsHowl = COMMON_BUILDER.comment("Whether higher age ranks buffs the mobs summoned by howling").define("ageBuffsHowl", true);
         bitingGivesFood = COMMON_BUILDER.comment("When enabled, biting an entity").define("bitingGivesFood", true);
         rankForBiteFood = COMMON_BUILDER.comment("Requires bitingGivesFood to be enabled. At what rank should biting begin to give food").defineInRange("rankForBitefood", 2, 0, 5);
